refactor: it 테스트는 패키지명을 명확화
